+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Change to requiring `family` options that agree with `glm`
 * Adding `family = gaussian_fixed_sd()` option to pre-specify a known standard deviation
+* Adding `family = lognormal()` option (e.g., for linearized Beverton-Holt stock recruit analysis)
 
 # dsem 2.0.0