+++ NEWS.md
@@ -1,4 +1,4 @@
-# dsem 1.7.0.9000
+# dsem 2.0.0
 
 * Add option to specify a path based fixed at another variable
 * Renamed `gmrf_parameterization = "conditional_krig"` as